Predict sales for each store-day level for one month.
Demand Forecast is one of the critical tasks in the Supply Chain and Retail Domain in general. It is key to the effective operation and optimization of the retail supply chain. Effectively solving this problem requires knowledge about a wide range of Data Sciences tricks and understanding ensemble techniques. 
To predict sales for each Store-Day level for one month. All the features will be provided and actual sales that happened during that month will also be provided for model evaluation.